requirements.txt
pandas==1.3.3
scikit-learn==0.24.2
dash==1.21.0
matplotlib==3.4.3
requests==2.26.0
joblib==1.0.1
pyyaml==5.4.1
